Improved PowerShell scripts
- Updated the error handling and flow on failure.
- Defaulted param scripts to have high verbosity.
- Updated readme with script requirements.
- Fixed deploy script to have the new solution name. I.e. ProjectPrefix+SolutionName

Fixes #23

+++ generators/app/templates/Nullfactory.Xrm.Tooling/Scripts/Pull-CrmSolution.ps1
@@ -30,17 +30,24 @@ $creds = New-Object System.Management.Automation.PSCredential ($username, $secur
 if($isOnPremServer)
 {
 	Write-Verbose "Connecting to the OnPrem crm instance $serverUrl ..."
-	Connect-CrmOnPremDiscovery -Credential $creds -ServerUrl $serverUrl
+	$connectionState = Connect-CrmOnPremDiscovery -Credential $creds -ServerUrl $serverUrl
 }
 else
 {
 	Write-Verbose "Connecting to the Online crm instance $serverUrl ..."
-	Connect-CrmOnline -Credential $creds -ServerUrl $serverUrl
+	$connectionState = Connect-CrmOnline -Credential $creds -ServerUrl $serverUrl
+}
+
+# Make sure that the connection was successful, else stop the script.
+if($connectionState.IsReady -eq $false)
+{
+	throw $connectionState.LastCrmError
 }
 
 $exportZipFileName = $solutionName + "_export.zip"
 $exportZipFileNameManaged = $solutionName + "_export_Managed.zip"
 
+# Clear previous instances of the exported solution file
 if (Test-Path $exportZipFileName) { Remove-Item $exportZipFileName }
 if (Test-Path $exportZipFileNameManaged) { Remove-Item $exportZipFileNameManaged }
 
@@ -53,7 +60,7 @@ Export-CrmSolution -SolutionName $solutionName -Managed -SolutionZipFileName $ex
 Write-Verbose "Delete the source controlled artifacts while keeping the project file intact..."
 Remove-Item -Recurse $solutionRootFolder  -Force -exclude *.csproj
 
-Write-Verbose "Extracing previous generated solution..."
+Write-Verbose "Unpacking exported solution..."
 if($solutionMapFile -eq "")
 {
 	.\..\bin\coretools\SolutionPackager.exe `
@@ -65,7 +72,7 @@ if($solutionMapFile -eq "")
 }
 else
 {
-	Write-Verbose "using mapping file $solutionMapFile for packaging"
+	Write-Verbose "Using mapping file $solutionMapFile for unpacking"
 	.\..\bin\coretools\SolutionPackager.exe `
 	/a:extract `
 	/packagetype:both `
@@ -73,4 +80,14 @@ else
 	/z:"$exportZipFileName" `
 	/ad:no `
 	/map:"$solutionMapFile"
+}
+
+# Verify that the solution packager ran successfully.
+if($?)
+{
+	Write-Verbose "Unpack successful!"
+}
+else
+{
+	throw "SolutionPackager unpack failed."
 }
